requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0
deep-sort-realtime>=1.3.2
streamlit>=1.30.0
pandas>=2.0.0
python-dotenv>=1.0.0
